+++ src/Config.php
@@ -3,6 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\OperatingSystem;
 
+use Innmind\TimeContinuum\{
+    Clock,
+    Earth,
+};
 use Innmind\Filesystem\CaseSensitivity;
 use Innmind\Server\Status\EnvironmentPath;
 use Innmind\Stream\{
@@ -13,6 +17,7 @@ use Innmind\Immutable\Maybe;
 
 final class Config
 {
+    private Clock $clock;
     private CaseSensitivity $caseSensitivity;
     private Capabilities $streamCapabilities;
     private EnvironmentPath $path;
@@ -23,11 +28,13 @@ final class Config
      * @param Maybe<positive-int> $maxHttpConcurrency
      */
     private function __construct(
+        Clock $clock,
         CaseSensitivity $caseSensitivity,
         Capabilities $streamCapabilities,
         EnvironmentPath $path,
         Maybe $maxHttpConcurrency,
     ) {
+        $this->clock = $clock;
         $this->caseSensitivity = $caseSensitivity;
         $this->streamCapabilities = $streamCapabilities;
         $this->path = $path;
@@ -40,6 +47,7 @@ final class Config
         $maxHttpConcurrency = Maybe::nothing();
 
         return new self(
+            new Earth\Clock,
             CaseSensitivity::sensitive,
             Streams::fromAmbientAuthority(),
             EnvironmentPath::of(\getenv('PATH') ?: ''),
@@ -50,9 +58,24 @@ final class Config
     /**
      * @psalm-mutation-free
      */
+    public function withClock(Clock $clock): self
+    {
+        return new self(
+            $clock,
+            $this->caseSensitivity,
+            $this->streamCapabilities,
+            $this->path,
+            $this->maxHttpConcurrency,
+        );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
     public function caseInsensitiveFilesystem(): self
     {
         return new self(
+            $this->clock,
             CaseSensitivity::insensitive,
             $this->streamCapabilities,
             $this->path,
@@ -66,6 +89,7 @@ final class Config
     public function useStreamCapabilities(Capabilities $capabilities): self
     {
         return new self(
+            $this->clock,
             $this->caseSensitivity,
             $capabilities,
             $this->path,
@@ -79,6 +103,7 @@ final class Config
     public function withEnvironmentPath(EnvironmentPath $path): self
     {
         return new self(
+            $this->clock,
             $this->caseSensitivity,
             $this->streamCapabilities,
             $path,
@@ -94,11 +119,20 @@ final class Config
     public function limitHttpConcurrencyTo(int $max): self
     {
         return new self(
+            $this->clock,
             $this->caseSensitivity,
             $this->streamCapabilities,
             $this->path,
             Maybe::just($max),
         );
+    }
+
+    /**
+     * @internal
+     */
+    public function clock(): Clock
+    {
+        return $this->clock;
     }
 
     /**