+++ src/Factory.php
@@ -10,14 +10,12 @@ use Innmind\TimeContinuum\{
 
 final class Factory
 {
-    public static function build(
-        Clock $clock = null,
-        Config $config = null,
-    ): OperatingSystem {
+    public static function build(Config $config = null): OperatingSystem
+    {
         switch (\PHP_OS) {
             case 'Darwin':
             case 'Linux':
-                return OperatingSystem\Unix::of($clock ?? new Earth\Clock, $config);
+                return OperatingSystem\Unix::of($config);
         }
 
         throw new \LogicException('Unuspported operating system');